Extract `StaticImportConfig` from `{StaticImport,NonStaticImport}`
Summary of changes:
- Collapse the `StaticImport:CandidateTypes`/`StaticImport:CandidateMembers`
  flags into a single `StaticImport:AdditionalCandidates` flag, mirroring
  `LexicographicalAnnotationAttributeListing`'s single-flag, `#`-disambiguated
  syntax.
- Extract the two checks' shared candidate-matching configuration and
  decision logic into a new `StaticImportConfig` class, injected into both
  checks rather than reached into across classes, following the shape of
  Error Prone's own `WellKnownKeep`.
- Extract `TypeMemberSpec`, a small shared parser for `<type>` and
  `<type>#<member>` flag entries, used by both `StaticImportConfig` and
  `AnnotationAttributeMatcher`.
