+++ error-prone-contrib/src/test/java/tech/picnic/errorprone/bugpatterns/NonStaticImportTest.java
@@ -1,41 +1,10 @@
 package tech.picnic.errorprone.bugpatterns;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import com.google.errorprone.BugCheckerRefactoringTestHelper;
 import com.google.errorprone.CompilationTestHelper;
 import org.junit.jupiter.api.Test;
 
 final class NonStaticImportTest {
-  @Test
-  void candidateTypesDoNotClash() {
-    assertThat(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_TYPES)
-        .doesNotContainAnyElementsOf(StaticImport.STATIC_IMPORT_CANDIDATE_TYPES);
-  }
-
-  @Test
-  void candidateMembersAreNotRedundant() {
-    assertThat(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_MEMBERS.keySet())
-        .doesNotContainAnyElementsOf(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_TYPES);
-
-    assertThat(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_MEMBERS.values())
-        .doesNotContainAnyElementsOf(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_IDENTIFIERS);
-  }
-
-  @Test
-  void candidateMembersDoNotClash() {
-    assertThat(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_MEMBERS.entries())
-        .doesNotContainAnyElementsOf(StaticImport.STATIC_IMPORT_CANDIDATE_MEMBERS.entries());
-  }
-
-  @SuppressWarnings(
-      "java:S3415" /* Comparing a constant against a non-constant value is intentional here. */)
-  @Test
-  void candidateIdentifiersDoNotClash() {
-    assertThat(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_IDENTIFIERS)
-        .doesNotContainAnyElementsOf(StaticImport.STATIC_IMPORT_CANDIDATE_MEMBERS.values());
-  }
-
   @Test
   void identification() {
     CompilationTestHelper.newInstance(NonStaticImport.class, getClass())
@@ -122,7 +91,7 @@ final class NonStaticImportTest {
   @Test
   void identificationWithAdditionalStaticImportCandidateMembers() {
     CompilationTestHelper.newInstance(NonStaticImport.class, getClass())
-        .setArgs("-XepOpt:StaticImport:CandidateMembers=reactor.core.publisher.Flux#just")
+        .setArgs("-XepOpt:StaticImport:AdditionalCandidates=reactor.core.publisher.Flux#just")
         .addSourceLines(
             "A.java",
             "import static reactor.core.publisher.Flux.just;",
