+++ error-prone-contrib/src/test/java/tech/picnic/errorprone/bugpatterns/StaticImportTest.java
@@ -1,101 +1,10 @@
 package tech.picnic.errorprone.bugpatterns;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.params.provider.Arguments.arguments;
-
-import com.google.common.collect.ImmutableMap;
 import com.google.errorprone.BugCheckerRefactoringTestHelper;
 import com.google.errorprone.CompilationTestHelper;
-import com.google.errorprone.ErrorProneFlags;
-import java.time.Clock;
-import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
 final class StaticImportTest {
-  @Test
-  void candidateTypesDoNotClash() {
-    assertThat(StaticImport.STATIC_IMPORT_CANDIDATE_TYPES)
-        .doesNotContainAnyElementsOf(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_TYPES);
-  }
-
-  @Test
-  void candidateMembersAreNotRedundant() {
-    assertThat(StaticImport.STATIC_IMPORT_CANDIDATE_MEMBERS.keySet())
-        .doesNotContainAnyElementsOf(StaticImport.STATIC_IMPORT_CANDIDATE_TYPES);
-  }
-
-  @Test
-  void candidateMembersDoNotClash() {
-    assertThat(StaticImport.STATIC_IMPORT_CANDIDATE_MEMBERS.entries())
-        .doesNotContainAnyElementsOf(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_MEMBERS.entries());
-
-    assertThat(StaticImport.STATIC_IMPORT_CANDIDATE_MEMBERS.values())
-        .doesNotContainAnyElementsOf(NonStaticImport.NON_STATIC_IMPORT_CANDIDATE_IDENTIFIERS);
-  }
-
-  private static Stream<Arguments> additionalCandidateValidationTestCases() {
-    /* { flag, value, expectedError } */
-    return Stream.of(
-        arguments(
-            "StaticImport:CandidateTypes",
-            Clock.class.getCanonicalName(),
-            "Invalid `StaticImport:CandidateTypes` flag value 'java.time.Clock': type is a "
-                + "non-static import candidate"),
-        arguments(
-            "StaticImport:CandidateTypes",
-            "java.lang.Math, java.util.UUID",
-            "Invalid `StaticImport:CandidateTypes` flag value ' java.util.UUID': expected a "
-                + "canonical type name"),
-        arguments(
-            "StaticImport:CandidateMembers",
-            Math.class.getCanonicalName(),
-            "Invalid `StaticImport:CandidateMembers` flag value 'java.lang.Math': expected format "
-                + "is '<type>#<member>'"),
-        arguments(
-            "StaticImport:CandidateMembers",
-            "java.lang.Math#min#max",
-            "Invalid `StaticImport:CandidateMembers` flag value 'java.lang.Math#min#max': expected "
-                + "format is '<type>#<member>'"),
-        arguments(
-            "StaticImport:CandidateMembers",
-            "#hash",
-            "Invalid `StaticImport:CandidateMembers` flag value '#hash': expected format is "
-                + "'<type>#<member>'"),
-        arguments(
-            "StaticImport:CandidateMembers",
-            "java.util.Objects#",
-            "Invalid `StaticImport:CandidateMembers` flag value 'java.util.Objects#': expected "
-                + "format is '<type>#<member>'"),
-        arguments(
-            "StaticImport:CandidateMembers",
-            "java.lang.Math#max, java.lang.Math#min",
-            "Invalid `StaticImport:CandidateMembers` flag value ' java.lang.Math#min': expected "
-                + "format is '<type>#<member>'"),
-        arguments(
-            "StaticImport:CandidateMembers",
-            "com.example.Foo#of",
-            "Invalid `StaticImport:CandidateMembers` flag value 'com.example.Foo#of': identifier "
-                + "is a non-static import candidate"),
-        arguments(
-            "StaticImport:CandidateMembers",
-            "java.util.Locale#ROOT",
-            "Invalid `StaticImport:CandidateMembers` flag value 'java.util.Locale#ROOT': member "
-                + "is a non-static import candidate"));
-  }
-
-  @MethodSource("additionalCandidateValidationTestCases")
-  @ParameterizedTest
-  void additionalCandidateValidation(String flag, String value, String expectedError) {
-    assertThatThrownBy(
-            () -> new StaticImport(ErrorProneFlags.fromMap(ImmutableMap.of(flag, value))))
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessage(expectedError);
-  }
-
   @Test
   void identification() {
     CompilationTestHelper.newInstance(StaticImport.class, getClass())
@@ -195,8 +104,8 @@ final class StaticImportTest {
   void identificationWithAdditionalCandidates() {
     CompilationTestHelper.newInstance(StaticImport.class, getClass())
         .setArgs(
-            "-XepOpt:StaticImport:CandidateTypes=java.lang.Math",
-            "-XepOpt:StaticImport:CandidateMembers=java.util.Objects#hash,reactor.core.publisher.Flux#just")
+            "-XepOpt:StaticImport:AdditionalCandidates=java.lang.Math,"
+                + "java.util.Objects#hash,reactor.core.publisher.Flux#just")
         .addSourceLines(
             "A.java",
             "import java.util.Objects;",
@@ -394,8 +303,8 @@ final class StaticImportTest {
   void replacementWithAdditionalCandidates() {
     BugCheckerRefactoringTestHelper.newInstance(StaticImport.class, getClass())
         .setArgs(
-            "-XepOpt:StaticImport:CandidateTypes=java.lang.Math",
-            "-XepOpt:StaticImport:CandidateMembers=reactor.core.publisher.Flux#just")
+            "-XepOpt:StaticImport:AdditionalCandidates=java.lang.Math,"
+                + "reactor.core.publisher.Flux#just")
         .addInputLines(
             "A.java",
             "import reactor.core.publisher.Flux;",
